requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
zipp==3.20.2
Flask-SQLAlchemy==3.1.1
redis==5.2.0
Flask-Cors==4.0.1